refactor: grouping common providers for App and Storybook
