+++ js/.storybook/preview.js
@@ -1,16 +1,8 @@
 import React from 'react'
-import { AppDimensionsProvider } from '@berty/contexts/app-dimensions.context'
-import { StyleProvider } from '@berty/contexts/styles'
-import { SafeAreaProvider } from 'react-native-safe-area-context'
-import { Provider as ReduxProvider } from 'react-redux'
-import { IconRegistry } from '@ui-kitten/components'
-import reduxStore from '@berty/redux/store'
-import { EvaIconsPack } from '@ui-kitten/eva-icons'
-import { FeatherIconsPack } from '../packages/messenger-app/feather-icons'
-import { CustomIconsPack } from '../packages/messenger-app/custom-icons'
 import { createNativeStackNavigator } from '@react-navigation/native-stack'
 import { NavigationContainer } from '@react-navigation/native'
 import { ScrollView, StyleSheet, View } from 'react-native'
+import AppCommonProviders from '@berty/contexts/AppCommonProviders'
 
 export const decorators = []
 export const parameters = {}
@@ -20,16 +12,7 @@ export const ScroolViewDecorator = (getStory, context) => (
 )
 
 export const AppDecorator = (getStory, context) => (
-	<SafeAreaProvider>
-		<AppDimensionsProvider>
-			<StyleProvider>
-				<ReduxProvider store={reduxStore}>
-					<IconRegistry icons={[EvaIconsPack, FeatherIconsPack, CustomIconsPack]} />
-					{getStory(context)}
-				</ReduxProvider>
-			</StyleProvider>
-		</AppDimensionsProvider>
-	</SafeAreaProvider>
+	<AppCommonProviders>{getStory(context)}</AppCommonProviders>
 )
 
 const StoryBookStack = createNativeStackNavigator()
