+++ js/packages/assets/custom-icons.tsx
