+++ js/packages/utils/testing/renderScreen.test.tsx
@@ -6,9 +6,9 @@ import React, { ComponentProps } from 'react'
 import { SafeAreaProvider } from 'react-native-safe-area-context'
 import { Provider } from 'react-redux'
 
+import { CustomIconsPack } from '@berty/assets/custom-icons'
+import { FeatherIconsPack } from '@berty/assets/feather-icons'
 import { UIKittenProvider } from '@berty/contexts/uiKitten.context'
-import { CustomIconsPack } from '@berty/messenger-app/custom-icons'
-import { FeatherIconsPack } from '@berty/messenger-app/feather-icons'
 import { isReadyRef, navigationRef } from '@berty/navigation/rootRef'
 import { ScreenFC, ScreensParams } from '@berty/navigation/types'
 import store from '@berty/redux/store'
